refactor: 기물 색깔 enum 클래스화
