+++ src/main/java/softeer2nd/domain/ChessColor.java
@@ -1,0 +1,6 @@
+package softeer2nd.domain;
+
+public enum ChessColor {
+    WHITE,
+    BLACK,
+}
