+++ src/main/java/softeer2nd/domain/Pawn.java
@@ -1,0 +1,13 @@
+package softeer2nd.domain;
+
+public class Pawn {
+    private final ChessColor color;
+
+    public Pawn(ChessColor color){
+        this.color=color;
+    }
+
+    public ChessColor getColor() {
+        return color;
+    }
+}
