--- src/main/java/softeer2nd/piece/Pawn.java
@@ -1,13 +1,0 @@
-package softeer2nd.piece;
-
-public class Pawn {
-    private final String color;
-
-    public Pawn(String color){
-        this.color=color;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
