+++ src/test/java/softeer2nd/domain/PawnTest.java
@@ -1,4 +1,4 @@
-package softeer2nd.piece;
+package softeer2nd.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class PawnTest {
-    private static final String WHITE = "white";
-    private static final String BLACK = "black";
+    private static final ChessColor WHITE = ChessColor.WHITE;
+    private static final ChessColor BLACK = ChessColor.BLACK;
 
     @Test
     @DisplayName("색깔을 가진 폰이 생성되어야 한다.")
@@ -16,7 +16,7 @@ public class PawnTest {
         verifyPawn(BLACK);
     }
 
-    void verifyPawn(final String color){
+    void verifyPawn(final ChessColor color){
         Pawn pawn = new Pawn(color);
         assertThat(pawn.getColor()).isEqualTo(color);
     }
